Fix warning

+++ src/sqrt.c
@@ -132,7 +132,7 @@ graph		--	The parent
 #define cury (*Cury)
     int i;
     if (graph->down[kid]->options) {
-        for (i = 0; i < strlen(graph->down[kid]->options); i++)
+        for (i = 0; i < (int)strlen(graph->down[kid]->options); i++)
             (*screen)[cury + graph->down[kid]->dim.baseline -
                       (graph->down[kid]->dim.y - 1) / 2 - 1][curx++] =
                 graph->down[kid]->options[i];
